solution/my_project_part_1/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
click==8.0.3
colorama==0.4.4
Django==4.0.1
django-nose==1.4.7
factory-boy==3.2.1
Faker==11.3.0
isort==5.10.1
marshmallow==3.14.1
mypy-extensions==0.4.3
nose==1.3.7
parameterized==0.8.1
pathspec==0.9.0
pinocchio==0.4.3
platformdirs==2.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.3
typing_extensions==4.0.1
